MEDI-93-Fullstack-Dev-Authentication-and-Authorisation

1. installed bcrypt and hashed the password
2. added login with hashedpassword in database
3. added JWT token generation to login and signup, storing privatekey in .env ,generated using node crypto function
4. created a new checkAuth middleware to protect routes on the server and implemented it in the product-requests-routes.js
5. Token set in the AuthContext, and used to send and receive data from the server. Fixed a few bugs
6. Added Server Side authorization to update a product request
7. Added server side authorization to delete a product
8. Added server side authorization to get all products requests by a user
9. updated useEffect hook dependency
10. added functionality that only the productRequestCreator can update the productRequest
11. added functionality that only productRequestCreator can Delete a productRequest
12. updated some error codes
13. stored token and userdata in localstorage
14. added auto-relogin for the user if the token is still in localstorage
15. remove token from localStorage when user logs out
16. added token expiry date to match token expiry from server(2hours)
17. added autologout of the user if token as expired
18. use userid from the token to create a productRequest ,rather than pass it in the body of the request

+++ frontend/src/productRequests/components/productRequestItem.jsx
@@ -30,7 +30,7 @@ const ProductRequestItem = (props) => {
   const handleConfirmDelete = async () => {
     setOpen(false);
     console.log("Deleting product request...");
-    const response =  await medishortsService.deleteProductRequest(props.id);
+    const response =  await medishortsService.deleteProductRequest(props.id , {headers: {Authorization: `Bearer ${auth.token}`}});
     props.onDelete(props.id)
     console.log(response);
     if (response.state !== true) {
